[feat] #66 어드민 페이지에서 특정 유저의 포인트 조회 API 기능 구현

+++ src/main/java/org/festimate/team/participant/repository/ParticipantRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface ParticipantRepository extends JpaRepository<Participant, Integer> {
+public interface ParticipantRepository extends JpaRepository<Participant, Long> {
     Participant getParticipantByUserAndFestival(User user, Festival festival);
 
     @Query("SELECT p FROM Participant p WHERE p.user = :user AND " +
